+++ packages/react-dom/types/Focus.d.ts
@@ -1,6 +1,7 @@
 import { ReactNode, Ref } from "react";
 export interface FocusProps {
     children(ref: Ref<any>): ReactNode;
+    preventScroll: boolean;
 }
 declare const Focus: (props: FocusProps) => JSX.Element;
 export default Focus;