+++ packages/react-dom/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Add `preventScroll` prop to `<Focus>`. When `false` (default), the page will scroll to the element that it focuses. When `true`, the page will not scroll.
+
 ## 1.0.0-beta.0
 
 * Added `@curi/react-dom`. This package has its own `<Link>` and `<Focus>` components, which rely on DOM APIs. This package re-exports all of `@curi/react`'s exports.